refactor findNodes xpath
