+++ src/nodes/loadNativeElements.ts
@@ -1,6 +1,7 @@
 import { extractTemplateExpressions } from "../html/evaluateTemplate";
 
 import { ReactiveNodesList } from "../types";
+import { findNodes } from "./findNodes";
 import { isCustomElement } from "./isCustomElement";
 import { registerReactiveNode } from "./registerReactiveNode";
 
@@ -8,26 +9,11 @@ export const registerNativeElements = (
     rootElement: Node,
     reactiveNodes: ReactiveNodesList
 ) => {
-    const elements: HTMLElement[] = [];
-    const xpath =
-        "self::*[text()[contains(., '{{')] and text()[contains(., '}}')]] | self::*[@*[contains(., '{{') and contains(., '}}')]] | .//*[text()[contains(., '{{')] and text()[contains(., '}}')]] | .//*[@*[contains(., '{{') and contains(., '}}')]]";
-
-    const result = document.evaluate(
-        xpath,
+    const elements = findNodes<HTMLElement>(
         rootElement,
-        null,
-        XPathResult.ORDERED_NODE_ITERATOR_TYPE,
-        null
+        "self::*[text()[contains(., '{{')] and text()[contains(., '}}')]] | self::*[@*[contains(., '{{') and contains(., '}}')]] | .//*[text()[contains(., '{{')] and text()[contains(., '}}')]] | .//*[@*[contains(., '{{') and contains(., '}}')]]",
+        (element) => !isCustomElement(element)
     );
-    let element = <HTMLElement>result.iterateNext();
-
-    while (element) {
-        if (!isCustomElement(element)) {
-            elements.push(element);
-        }
-
-        element = <HTMLElement>result.iterateNext();
-    }
 
     for (let i = 0; i < elements.length; i++) {
         const elementId = reactiveNodes.id();
